[Refactor] : 메인 js 코드를 공통 js 파일로 옮김
